requirements.txt
streamlit
pdfplumber
python-docx
rapidfuzz
pandas
numpy
openai
langchain
faiss-cpu
